Revert "Merge pull request #11 from mr12n21/completing-app-v1"
This reverts commit 7a9185a87df02b01b8a643bf245075b7eb9a372f, reversing
changes made to 80c06f190f7bb2dbe88e415a11b25d70ff3af839.
