+++ .gitignore
@@ -2,6 +2,5 @@
 /bin
 /lib
 /lib64
-/include
 /data-service/output-labels
 pyvenv.cfg
