+++ data-service/README.md
@@ -1,0 +1,21 @@
+co je jeste treba resit:
+
+vypocet karavan
+
+mezi datum dat / misto . a datum co nejvi zkratit aby zabral co nejmene mista
+
+trocu zmensit cely stitek aby se dal nalepit na cedulku
+
+
+celkova kompletace na rpi ktere bude zajistovat veskere sluby po ulozeni do smb
+
+    udelat system ktery bude archivovat vsechna pdf aby byli dostupna v pripade problemu a pro zpetnou analizu
+    
+    /data
+        /pdf    -ulozeni pdf prevod na tisknutelne png
+        /archiv
+
+
+###
+
+udelat js rozsireni ktere bude zjednodusovat ukladani na smb-share pdf
