+++ requirements.txt
@@ -1,24 +1,5 @@
-attrs==24.3.0
-blinker==1.9.0
-brother-ql==0.9.4
-cffi==1.17.1
-chardet==5.2.0
-charset-normalizer==3.4.1
-click==8.1.8
-cryptography==44.0.0
-Flask==3.1.0
-future==1.0.0
-itsdangerous==2.2.0
-Jinja2==3.1.5
-MarkupSafe==3.0.2
-packbits==0.6
-pdfminer.six==20231228
-pdfplumber==0.11.4
-pillow==11.0.0
-pycparser==2.22
-pypdfium2==4.30.1
-pyusb==1.2.1
-PyYAML==6.0.2
-reportlab==4.2.5
-watchdog==6.0.0
-Werkzeug==3.1.3
+reportlab
+brother_ql
+Pillow
+pdfplumber
+PyYAML
